Servicios

+++ ED-Backend/src/main/java/com/easydept/edbackend/repositories/EdificiosRepository.java
@@ -3,5 +3,5 @@ package com.easydept.edbackend.repositories;
 import com.easydept.edbackend.entity.Edificio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EdificiosRepository extends JpaRepository<Edificio, Long> {
+public interface EdificiosRepository extends JpaRepository<Edificio, Integer> {
 }
